+++ src/main/java/Stitching_2D.java
@@ -953,8 +953,8 @@ public class Stitching_2D implements PlugIn
 			return result;
 		}
 
-		avg1 /= (double) count;
-		avg2 /= (double) count;
+		avg1 /= count;
+		avg2 /= count;
 
 		double var1 = 0, var2 = 0;
 		double coVar = 0;
@@ -991,10 +991,10 @@ public class Stitching_2D implements PlugIn
 				}
 			}
 
-		SSQ /= (double) count;
-		var1 /= (double) count;
-		var2 /= (double) count;
-		coVar /= (double) count;
+		SSQ /= count;
+		var1 /= count;
+		var2 /= count;
+		coVar /= count;
 
 		double stDev1 = Math.sqrt(var1);
 		double stDev2 = Math.sqrt(var2);
@@ -1332,7 +1332,7 @@ public class Stitching_2D implements PlugIn
 
 			for (int y = 0; y < height; y++)
 				for (int x = 0; x < width; x++)
-					pixels.data[pixels.getPos(x, y)] = (float) (pixelTmp[count++] & 0xff);
+					pixels.data[pixels.getPos(x, y)] = pixelTmp[count++] & 0xff;
 		}
 		else if (ip.getPixels() instanceof short[])
 		{
@@ -1341,7 +1341,7 @@ public class Stitching_2D implements PlugIn
 
 			for (int y = 0; y < height; y++)
 				for (int x = 0; x < width; x++)
-					pixels.data[pixels.getPos(x, y)] = (float) (pixelTmp[count++] & 0xffff);
+					pixels.data[pixels.getPos(x, y)] = pixelTmp[count++] & 0xffff;
 		}
 		else if (ip.getPixels() instanceof int[]) 
 		{
@@ -1408,7 +1408,7 @@ public class Stitching_2D implements PlugIn
 
 			for (int y = yCrop; y < yCrop + hCrop; y++)
 				for (int x = xCrop; x < xCrop + wCrop; x++)
-					pixels.data[pixels.getPos(x - xCrop, y - yCrop)] = (float) (pixelTmp[x + y * width] & 0xff);
+					pixels.data[pixels.getPos(x - xCrop, y - yCrop)] = pixelTmp[x + y * width] & 0xff;
 		}
 		else if (ip.getPixels() instanceof short[])
 		{
@@ -1416,7 +1416,7 @@ public class Stitching_2D implements PlugIn
 
 			for (int y = yCrop; y < yCrop + hCrop; y++)
 				for (int x = xCrop; x < xCrop + wCrop; x++)
-					pixels.data[pixels.getPos(x - xCrop, y - yCrop)] = (float) (pixelTmp[x + y * width] & 0xffff);
+					pixels.data[pixels.getPos(x - xCrop, y - yCrop)] = pixelTmp[x + y * width] & 0xffff;
 		}
 		else if (ip.getPixels() instanceof int[])
 		{
